% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golub99.R
\docType{data}
\name{golub99}
\alias{golub99}
\title{golub99 data}
\format{
a list
}
\usage{
data(golub99)
}
\value{
golub99 data, a list containing 5147 genes and binary response.
}
\description{
The golub99 dataset, derived from a leukemia study, is probably the most
well-known gene expression cancer dataset. golub99 data have 72 samples
and 5147 genes. The data has binary outcomes.
}
\examples{
data(golub99)
}
\references{
Golub, Todd R.; Slonim, Donna K.; Tamayo, Pablo; Huard, Christine;
Gaasenbeek, Michelle; Mesirov, Jill P. et al. (1999): Molecular classification
of cancer: class discovery and class prediction by gene expression monitoring.
In science 286 (5439), pp. 531–537.
}
\keyword{datasets}
