% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Path.R
\name{plot_Path}
\alias{plot_Path}
\title{Visualize Influential Scores Across Boosting Iterations}
\usage{
plot_Path(
  obj,
  aims = c("variableSelection", "prediction"),
  ref = c("mean", "orig."),
  topN = 3
)
}
\arguments{
\item{obj}{An object resulting from a leave-one-out analysis performed
by \code{glmboostDrop1}. Specifically, \code{obj} should be a list containing \code{LooObj},
which holds the leave-one-out objects.}

\item{aims}{A character vector indicating the aim of the influential
observation detection: "variableSelection" for IPs detection for the aim of
variable selection, "prediction" for the aim of prediction, or both.}

\item{ref}{A character string specifying the reference for comparison. Valid
options are "mean" for the average similarity across all leave-one-out
iterations, or "orig" for the similarity compared to the original model's
selected variables. This parameter influences how the similarity scores are
calculated and displayed in the plot.}

\item{topN}{An integer specifying the number of topN to emphasize as the
most influential, based on their distance from the reference. Default is 3.
These paths are highlighted in the plot, facilitating the identification of
observations with the most influence on either variable selection and
prediction.}
}
\description{
This function visualizes the influential score changes across different
boosting iterations. It depicts the influence of excluding each observation
on variable selection and prediction.
}
\details{
The plot generated by this function displays the number of boosting
iterations on the x-axis and the influential scores on the y-axis. Each line
represents one leave-out observation, illustrating the influence of a case on
either variable selection and prediction evolves with additional boosting
iterations. The \code{topN} most influential observations are emphasized in a
different color, and the selected variables at their optimal boosting
iteration are marked with a point. The path of the reference observation is
depicted with a dashed vertical line.
}
\examples{
library(mboost)
data(golub99)
X <- golub99$X
X <- scale(X)
y <- golub99$y
set.seed(1) 
foldid <- sample(rep(1:10, length = nrow(X)), nrow(X)) 
cv <- sapply(1:max(foldid), function(x) as.numeric(foldid != x))
obj <- mboost::glmboost(X, y,
                family = Binomial(),
                control = boost_control(mstop = 140,
                                        nu = 0.1,
                                        risk = "inbag"), 
                                        center = FALSE)
drop1obj <- glmboostDrop1(obj,
                          nCores = 1,
                          fixMstop = NULL,
                          aims = c("variableSelection", "prediction"),
                          ref = "mean",
                          folds = cv)
plot_Scores(drop1obj)
plot_Path(drop1obj, ref = "mean")
}
